OpenACC backend PR fixes
Fix ExampleDefaultAcc for OAcc -> alpaka::CtxThreadOacc<TDim,TIdx>

Fix TestAccs: Rm double AccOacc

Oacc Atomic,Sync: Fix guards

IdxBtOacc: rename, fix comment

Oacc: clean-up comments

Oacc: Buf,Copy: clean-up code

Oacc: Fix ci.yml

Oacc: merge CtxThreadOacc into AccOacc

AccOacc: Add final

dtor can never be virtual because of use in device code

Oacc::Copy: remove cast on compare to 0.

oacc: Remove /* virtual */ dtor comments

Oacc: rename IdxGbOacc

Oacc: clean-up code, comments

AtomicTestKernel::operator(): Rm TTAcc

README_OACC: Rm -fopenacc/-acc (set by CMake)

Mv IdxBtOacc -> IdxBtLinear

Oacc: Rm *::BlockShared

Oacc/Set: Add debug guard around print

BufOacc: Fix size DBG output

BufOmp5: Fix size DBG output

check OpenACC 2.0, error below
