+++ include/alpaka/block/sync/BlockSyncBarrierOacc.hpp
@@ -1,4 +1,4 @@
-/* Copyright 2019 Axel Huebl, Benjamin Worpitz
+/* Copyright 2020 Jeffrey Kelling
  *
  * This file is part of Alpaka.
  *
@@ -9,7 +9,11 @@
 
 #pragma once
 
-#ifdef _OPENACC
+#ifdef ALPAKA_ACC_ANY_BT_OACC_ENABLED
+
+#if _OPENACC < 201306
+    #error If ALPAKA_ACC_ANY_BT_OACC_ENABLED is set, the compiler has to support OpenACC 2.0 or higher!
+#endif
 
 #include <alpaka/block/sync/Traits.hpp>
 
@@ -17,13 +21,11 @@ namespace alpaka
 {
     //#############################################################################
     //! The OpenACC barrier block synchronization.
-    //! Traits are specialized on BlockSyncBarrierOaccBlockShared
     class BlockSyncBarrierOacc
     {
     public:
         //-----------------------------------------------------------------------------
-        ALPAKA_FN_HOST BlockSyncBarrierOacc()
-        {}
+        ALPAKA_FN_HOST BlockSyncBarrierOacc() = default;
         //-----------------------------------------------------------------------------
         ALPAKA_FN_HOST BlockSyncBarrierOacc(BlockSyncBarrierOacc const &) = delete;
         //-----------------------------------------------------------------------------
@@ -33,31 +35,14 @@ namespace alpaka
         //-----------------------------------------------------------------------------
         ALPAKA_FN_HOST auto operator=(BlockSyncBarrierOacc &&) -> BlockSyncBarrierOacc & = delete;
         //-----------------------------------------------------------------------------
-        /*virtual*/ ~BlockSyncBarrierOacc() = default;
+        ~BlockSyncBarrierOacc() = default;
 
-        class BlockShared : public concepts::Implements<ConceptBlockSync, BlockShared>
-        {
-            public:
-                //-----------------------------------------------------------------------------
-                ALPAKA_FN_HOST BlockShared()
-                {}
-                //-----------------------------------------------------------------------------
-                ALPAKA_FN_HOST BlockShared(BlockShared const &) = delete;
-                //-----------------------------------------------------------------------------
-                ALPAKA_FN_HOST BlockShared(BlockShared &&) = delete;
-                //-----------------------------------------------------------------------------
-                ALPAKA_FN_HOST auto operator=(BlockShared const &) -> BlockShared & = delete;
-                //-----------------------------------------------------------------------------
-                ALPAKA_FN_HOST auto operator=(BlockShared &&) -> BlockShared & = delete;
-                //-----------------------------------------------------------------------------
-                /*virtual*/ ~BlockShared() = default;
-
-                std::uint8_t mutable m_generation = 0u;
-                // NVHPC 20.7: initializer causes warning:
-                // NVC++-W-0155-External and Static variables are not supported in acc routine - _T139951818207704_37530
-                int mutable m_syncCounter[4] {0,0,0,0};
-                int mutable m_result;
-        };
+        std::uint8_t mutable m_generation = 0u;
+        // NVHPC 20.7: initializer causes warning:
+        // NVC++-W-0155-External and Static variables are not supported in acc routine - _T139951818207704_37530
+        //! m_synchCounter[ 2 generations  * 2 counters per]
+        int mutable m_syncCounter[2*2] {0,0,0,0};
+        int mutable m_result;
     };
 }
