+++ include/alpaka/mem/buf/oacc/Set.hpp
@@ -12,7 +12,7 @@
 #ifdef ALPAKA_ACC_ANY_BT_OACC_ENABLED
 
 #if _OPENACC < 201306
-    #error If ALPAKA_ACC_ANY_BT_OACC_ENABLED is set, the compiler has to support OpenACC xx or higher!
+    #error If ALPAKA_ACC_ANY_BT_OACC_ENABLED is set, the compiler has to support OpenACC 2.0 or higher!
 #endif
 
 #include <alpaka/dev/DevOacc.hpp>
@@ -79,7 +79,9 @@ namespace alpaka
                             pitch
                         ); // NOP if size is zero
 
+#if ALPAKA_DEBUG >= ALPAKA_DEBUG_MINIMAL
                 std::cout << "Set TDim=" << TDim::value << " pitch=" << pitch << " byteExtent=" << byteExtent << std::endl;
+#endif
                 auto elementsPerThread = Vec<TDim, Idx>::all(static_cast<Idx>(1u));
                 elementsPerThread[lastDim] = 4;
                 // Let alpaka calculate good block and grid sizes given our full problem extent
