+++ test/common/include/alpaka/test/acc/TestAccs.hpp
@@ -147,29 +147,7 @@ namespace alpaka
             template<
                 typename TDim,
                 typename TIdx>
-            using AccOaccIfAvailableElseInt = alpaka::CtxThreadOacc<TDim, TIdx>;
-#else
-            template<
-                typename TDim,
-                typename TIdx>
-            using AccOaccIfAvailableElseInt = int;
-#endif
-#if defined(ALPAKA_ACC_ANY_BT_OACC_ENABLED) && !(defined(TEST_UNIT_KERNEL_KERNEL_STD_FUNCTION)) \
-    && !( BOOST_COMP_GNUC && ( \
-            /* tests excluded because of GCC10 Oacc / Omp5 target symbol bug with multiple units */ \
-            defined(TEST_UNIT_BLOCK_SHARED) \
-            || defined(TEST_UNIT_BLOCK_SYNC) \
-            || defined(TEST_UNIT_WARP) \
-            || defined(TEST_UNIT_INTRINSIC) \
-            || defined(TEST_UNIT_KERNEL) \
-            || defined(TEST_UNIT_MEM_VIEW) \
-            || defined(TEST_UNIT_MATH) /* because of static const members */ \
-            || defined(TEST_UNIT_MEM_BUF) /* actually works, but hangs when ran by ctest */ \
-        ))
-            template<
-                typename TDim,
-                typename TIdx>
-            using AccOaccIfAvailableElseInt = alpaka::CtxThreadOacc<TDim, TIdx>;
+            using AccOaccIfAvailableElseInt = alpaka::AccOacc<TDim, TIdx>;
 #else
             template<
                 typename TDim,
